+++ lib/models/database_service.dart
@@ -1,23 +1,19 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 class DatabaseService {
-  final String? uid;
-  DatabaseService({this.uid});
+  final String? email;
+  DatabaseService({this.email});
 
   // reference for our collections
   final CollectionReference userCollection =
       FirebaseFirestore.instance.collection("users");
-  final CollectionReference groupCollection =
-      FirebaseFirestore.instance.collection("groups");
 
   // saving the userdata
-  Future savingUserData(String fullName, String email) async {
-    return await userCollection.doc(uid).set({
-      "fullName": fullName,
+  Future savingUserData(String userName, String email) async {
+    return await userCollection.doc(email).set({
+      "userName": userName,
       "email": email,
-      "groups": [],
       "profilePic": "",
-      "uid": uid,
     });
   }
 
@@ -28,106 +24,18 @@ class DatabaseService {
     return snapshot;
   }
 
-  // get user groups
-  getUserGroups() async {
-    return userCollection.doc(uid).snapshots();
-  }
-
-  // creating a group
-  Future createGroup(String userName, String id, String groupName) async {
-    DocumentReference groupDocumentReference = await groupCollection.add({
-      "groupName": groupName,
-      "groupIcon": "",
-      "admin": "${id}_$userName",
-      "members": [],
-      "groupId": "",
-      "recentMessage": "",
-      "recentMessageSender": "",
-    });
-    // update the members
-    await groupDocumentReference.update({
-      "members": FieldValue.arrayUnion(["${uid}_$userName"]),
-      "groupId": groupDocumentReference.id,
-    });
-
-    DocumentReference userDocumentReference = userCollection.doc(uid);
-    return await userDocumentReference.update({
-      "groups":
-          FieldValue.arrayUnion(["${groupDocumentReference.id}_$groupName"])
-    });
-  }
-
-  // getting the chats
-  getChats(String groupId) async {
-    return groupCollection
-        .doc(groupId)
+  getChats(String email) async {
+    return userCollection
+        .doc(email)
         .collection("messages")
         .orderBy("time")
         .snapshots();
   }
 
-  Future getGroupAdmin(String groupId) async {
-    DocumentReference d = groupCollection.doc(groupId);
-    DocumentSnapshot documentSnapshot = await d.get();
-    return documentSnapshot['admin'];
-  }
-
-  // get group members
-  getGroupMembers(groupId) async {
-    return groupCollection.doc(groupId).snapshots();
-  }
-
-  // search
-  searchByName(String groupName) {
-    return groupCollection.where("groupName", isEqualTo: groupName).get();
-  }
-
-  // function -> bool
-  Future<bool> isUserJoined(
-      String groupName, String groupId, String userName) async {
-    DocumentReference userDocumentReference = userCollection.doc(uid);
-    DocumentSnapshot documentSnapshot = await userDocumentReference.get();
-
-    List<dynamic> groups = await documentSnapshot['groups'];
-    if (groups.contains("${groupId}_$groupName")) {
-      return true;
-    } else {
-      return false;
-    }
-  }
-
-  // toggling the group join/exit
-  Future toggleGroupJoin(
-      String groupId, String userName, String groupName) async {
-    // doc reference
-    DocumentReference userDocumentReference = userCollection.doc(uid);
-    DocumentReference groupDocumentReference = groupCollection.doc(groupId);
-
-    DocumentSnapshot documentSnapshot = await userDocumentReference.get();
-    List<dynamic> groups = await documentSnapshot['groups'];
-
-    // if user has our groups -> then remove then or also in other part re join
-    if (groups.contains("${groupId}_$groupName")) {
-      await userDocumentReference.update({
-        "groups": FieldValue.arrayRemove(["${groupId}_$groupName"])
-      });
-      await groupDocumentReference.update({
-        "members": FieldValue.arrayRemove(["${uid}_$userName"])
-      });
-    } else {
-      await userDocumentReference.update({
-        "groups": FieldValue.arrayUnion(["${groupId}_$groupName"])
-      });
-      await groupDocumentReference.update({
-        "members": FieldValue.arrayUnion(["${uid}_$userName"])
-      });
-    }
-  }
-
   // send message
-  sendMessage(String groupId, Map<String, dynamic> chatMessageData) async {
-    groupCollection.doc(groupId).collection("messages").add(chatMessageData);
-    groupCollection.doc(groupId).update({
+  sendMessage(String email, Map<String, dynamic> chatMessageData) async {
+    userCollection.doc(email).collection("messages").add(chatMessageData);
+    userCollection.doc(email).update({
       "recentMessage": chatMessageData['message'],
       "recentMessageSender": chatMessageData['sender'],
       "recentMessageTime": chatMessageData['time'].toString(),
